Submission

+++ src/main/java/com/stackroute/muzixapp/service/TrackServiceImpl.java
@@ -47,8 +47,13 @@ public class TrackServiceImpl implements TrackService {
 
     @Override
     public void deleteTrack(int id) {
-
-        trackRepository.deleteById(id);
+        Optional<Track> track = trackRepository.findById(id);
+        if(track.isPresent()) {
+            trackRepository.deleteById(id);
+            return track.get();
+        } 
+        else {
+             return null;
 
     }
